+++ src/main/java/com/healthcare/system/services/implementation/HealthProviderServiceImpl.java
@@ -1,11 +1,15 @@
 package com.healthcare.system.services.implementation;
 
 import com.healthcare.system.entities.*;
+import com.healthcare.system.entities.Complaint;
+import com.healthcare.system.entities.HealthProvider;
+import com.healthcare.system.entities.Patient;
 import com.healthcare.system.exceptions.AlreadyLoggedInException;
 import com.healthcare.system.exceptions.AlreadyLoggedOutException;
 import com.healthcare.system.exceptions.ValidationException;
 import com.healthcare.system.exceptions.WrongCredentials;
-import com.healthcare.system.repositories.*;
+
+import com.healthcare.system.repositories.HealthProviderRepository;
 import com.healthcare.system.services.HealthProviderService;
 import com.healthcare.system.session.SessionManager;
 
@@ -23,22 +27,30 @@ public class HealthProviderServiceImpl implements HealthProviderService {
         this.healthProviderRepository = healthProviderRepository;
     }
     @Override
-    public void save(HealthProvider healthProvider) {
+    public void save(HealthProvider healthProvider) throws ValidationException {
+        verifyCredentials(HealthProvider.class,healthProvider);
         healthProviderRepository.save(healthProvider);
     }
 
     @Override
-    public HealthProvider getById(int id) {
+    public HealthProvider getById(int id) throws WrongCredentials {
+        if(healthProviderRepository.getById(id) == null) {
+            throw new WrongCredentials("HealthProvider with id: " + id + " does not exist");
+        }
         return healthProviderRepository.getById(id);
     }
 
     @Override
-    public HealthProvider deleteById(int id) {
+    public HealthProvider deleteById(int id) throws WrongCredentials {
+        if(healthProviderRepository.getById(id) == null) {
+            throw new WrongCredentials("HealthProvider with id: " + id + " does not exist");
+        }
         return healthProviderRepository.deleteById(id);
     }
 
     @Override
-    public void update(HealthProvider healthProvider) {
+    public void update(HealthProvider healthProvider) throws ValidationException {
+        verifyCredentials(HealthProvider.class,healthProvider);
         healthProviderRepository.update(healthProvider);
     }
 
@@ -115,9 +127,11 @@ public class HealthProviderServiceImpl implements HealthProviderService {
     }
 
     @Override
-    public List<Appointment> getAppointments(int healthProviderId) {
-        //@AnshMakker complete this method
-        return null;
+    public List<Appointment> getAppointments(int healthProviderId) throws WrongCredentials {
+        if(healthProviderRepository.getById(healthProviderId) == null) {
+            throw new WrongCredentials("HealthProvider with id: " + healthProviderId + " does not exist");
+        }
+        return healthProviderRepository.getById(healthProviderId).getAppointmentList();
     }
 
     @Override
