+++ src/main/java/com/healthcare/system/services/implementation/ReasonServiceImpl.java
@@ -1,8 +1,12 @@
 package com.healthcare.system.services.implementation;
 
 import com.healthcare.system.entities.Reason;
+import com.healthcare.system.exceptions.ValidationException;
+import com.healthcare.system.exceptions.WrongCredentials;
 import com.healthcare.system.repositories.ReasonRepository;
 import com.healthcare.system.services.ReasonService;
+
+import static com.healthcare.system.util.Verification.verifyCredentials;
 
 public class ReasonServiceImpl implements ReasonService {
 
@@ -13,22 +17,30 @@ public class ReasonServiceImpl implements ReasonService {
     }
 
     @Override
-    public void save(Reason reason) {
+    public void save(Reason reason) throws ValidationException {
+        verifyCredentials(Reason.class,reason);
         reasonRepository.save(reason);
     }
 
     @Override
-    public Reason findById(int id) {
+    public Reason findById(int id) throws WrongCredentials {
+        if(reasonRepository.findById(id) == null) {
+            throw new WrongCredentials("Reason with id: " + id + " does not exist");
+        }
         return reasonRepository.findById(id);
     }
 
     @Override
-    public void deleteById(int id) {
+    public void deleteById(int id) throws WrongCredentials {
+        if(reasonRepository.findById(id) == null) {
+            throw new WrongCredentials("Reason with id: " + id + " does not exist");
+        }
         reasonRepository.deleteById(id);
     }
 
     @Override
-    public void update(Reason reason) {
+    public void update(Reason reason) throws ValidationException {
+        verifyCredentials(Reason.class,reason);
         reasonRepository.update(reason);
     }
 }
